requirements.txt
annotated-types==0.6.0
anyio==4.2.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
fastapi==0.109.2
h11==0.14.0
idna==3.6
mysql-connector-python==8.3.0
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
pymongo==4.6.3
sniffio==1.3.0
starlette==0.36.3
typing_extensions==4.9.0
uvicorn==0.23.2
email-validator
dotenv
